2-11/12-20 Work
Implemented a bug fix for the double jumping error. Jumping logic now checks if any colliders tagged with Environment are touching the entity to determine if they are grounded.
Implemented UI dialogue into MVP_Level_1. Currently waiting for implementation of multiple actors in dialogue.
Implemented a visual color flash for enemies that are hit by the player. Functions off of a coroutine that briefly applies a color over the sprite of an Enemy entities sprite. The same logic could apply to all entities, and should in the future be assigned to the Character Controller as a general entity class.
Changed the enemy targeting to apply a Vector3 as the distance between the enemy and player. The enemy will determine the player's position, and approach that plus the new Vector. More complex dodging may be necessary at high level implementations to create "smarter" AI.
I genuinely can't remember if I added anything else, it's 2AM and I want to play Pokemon. This may be me talking to myself in a roundabout fashion.